Add Subversive m2e feature in a consistent way back
subversive.b3aggrcon referenced a feature group IU that lead to
inconsistencies in the b3 model. As a quick fix this reference has been
removed in commit 3b0b4100e2abe2fec806176582797784569e7f74. This new
change adds the same feature in a consistent way to the b3 model again.

Change-Id: I911d8e4834b0419aac176c3f985f0a9436d2fa5c
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>