Remove versionRange since we were not updating it
Currently, we never update the versionRange and the versions shown are
obsolete.

This was working because if we don't surround the version with [] it
allows for any version greater than the specified one.

However, having the version there is confusing and does not provide any
advantage in my eyes.

We could put the correct versions, but then at every milestone we would
have to update this file, which I'd like to avoid.

Instead, let's remove the versionRange completely to avoid confusion.
Then the latest version available int the repo should be automatically
selected.

When testing, we already check that the proper versions are part of the
build, therefore, we
are covered if something unexpected were to happen with the simrel
build.

Change-Id: I774a787bd66607e2dea803834109d1881bc4e09e