Bug 506502 - a pom.xml orchestrating aggregation and reports
This can be used as a replacement for some Ant scripts. Run `mvn clean
verify` and that will generate site according to b3 model and p2reports.
A first advantage is that this `mvn clean verify` commnd is simple to
run and has only Maven as requirements, so it's easer to use and
document, and user can even guess when they see a pom.xml which command
to run without much pre-requisite knowledge of the project

The provisioning of the builder app is provided via the eclipserun tycho
plugin, instead of Ant scripts.
This is a second advantage as it reduces the responsibilities of SimRel,
by reusing features provided by Tycho.

The output of `mvn verify` here is expected to be the same as with
current Ant-based orchestration, as it's in the end calling the same
application on the same model.
Also, as it only consists in addition of a file, the Ant-based
orchestration is still supposed to work with that change.

Change-Id: I3a6c67fe10982439c5ecb9af21589dad14ae904b
Signed-off-by: Mickael Istria <mistria@redhat.com>