Revert GMF Runtime to 1.10.0
This reverts commit c74d1853987934739db4778efc679955cba6669d. We're
quite late in the Neon.3 release cycle, which will be the last Neon
release. Even though the code changes between 1.10.0 and 1.10.1 are
small, I am not 100% confident that I have not missed anything in the
GMF Runtime build & release processes (as I took over the project's
maintenanceship recently).

GMF Runtime 1.10.1 will be released, but not contributed in Neon.3.
The single fix included in this version does not seem worth the
risk (however remote) of breaking GMF in the last ever Neon release.

Change-Id: I24b1b5c0d2ec5c852ca49d9b2f6e889600fd6709
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>