[releng] change default 'buildRoot' property value
... to 'buildresultsPhoton'

See also wiki at

https://wiki.eclipse.org/SimRel/Simultaneous_Release_Engineering#SimRel

which I use update to include

<quote>
In the simrel.aggrcon file itself, there is a property, named
'buildRoot' (or 'Build Root' if using the property page GUI from
Aggregation Editor) that should be update every major release to
something like 'buildresultsPhoton' or similar. The "buildresult" part
since the .gitignore file says to ignore '/buildresult*' files and the
last part, such as release name 'Photon' so that there can be different
areas for the current stream and the update stream.
</quote>


Change-Id: I2c58463bd9cd77443e11ea20f5d07d657d7f5fe0
Signed-off-by: David Williams <david_williams@acm.org>