CDT 9.11.0 M2, LSP4J 0.9.0 M2, LSP4E 0.14 M2
Because of API changes in LSP4J we need to update everything in one go.

The tm.terminal is now part of the CDT p2 repo rather than in a separate one.

Meson has been removed because it is not working properly with Java 8
and the required jackson versions are missing. See Bug 559641

Also, fix misformatted scout.aggrcon, result of open and saving
of aggr file

Change-Id: I354fd806080d3f9e3bb4352cdc8de9de8893d1c3