Bug 568003: Temporarily disable Mylyn's SDK contribution
Mylyn's SDK feature depends on lucene.core 7.1.0, but that is no longer
part of Orbit. The AERI project was contributing it, so Mylyn was
picking it up from there. However as AERI is being removed from SimRel
Mylyn needs to either update to newer version or provide 7.1.0 itself
to SimRel.

Change-Id: Idc5c17f9ead21b8ec8305b1ab142aa1e79f30442