Remove explicit bundles from CDT's contribution
These bundles have been manually listed for a long time (since first
version of these files in 2012). However these bundles end up in the
output even if not manually listed, therefore remove them being
explicit.

The secondary issue with this explicit list is that there are other
native bundles not listed here, so this commit also resolves this
inconsistency.